% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws.R
\name{aws}
\alias{aws}
\title{AWS downloader}
\usage{
aws(
  input_url,
  output_path = file.path(tempdir(), basename(input_url)),
  bucket,
  force_overwrite = FALSE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{input_url}{URL to remote file.}

\item{output_path}{The file name you want to save the download as.}

\item{bucket}{Character string with the name of the bucket, or an object of class \dQuote{s3_bucket}.}

\item{force_overwrite}{Overwrite existing file.}

\item{verbose}{Print messages.}

\item{...}{
  Arguments passed on to \code{\link[aws.s3:get_object]{aws.s3::save_object}}
  \describe{
    \item{\code{object}}{Character string with the object key, or an object of class \dQuote{s3_object}. In most cases, if \code{object} is specified as the latter, \code{bucket} can be omitted because the bucket name will be extracted from \dQuote{Bucket} slot in \code{object}.}
    \item{\code{headers}}{List of request headers for the REST call.}
    \item{\code{file}}{An R connection, or file name specifying the local file to save the object into.}
    \item{\code{overwrite}}{A logical indicating whether to overwrite \code{file}. Passed to \code{\link[httr]{write_disk}}. Default is \code{TRUE}.}
  }}
}
\value{
Path where the file has been downloaded
}
\description{
AWS S3 bucket downloader.
}
\examples{
bucket <- "s3://broad-alkesgroup-ukbb-ld/"
input_url <- "UKBB_LD/chr10_135000001_138000001.gz"
out <- aws(input_url = input_url, bucket = bucket)
}
\seealso{
\url{https://blog.djnavarro.net/posts/2022-03-17_using-aws-s3-in-r/}

Other downloaders: 
\code{\link{axel}()},
\code{\link{downloader}()},
\code{\link{wget}()}
}
\concept{downloaders}
