% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_vcf.R
\name{download_vcf}
\alias{download_vcf}
\title{Download a remote VCF file and its index file}
\usage{
download_vcf(
  vcf_url,
  vcf_dir = tempdir(),
  vcf_download = TRUE,
  download_method = "download.file",
  force_new = FALSE,
  quiet = TRUE,
  nThread = 1
)
}
\arguments{
\item{vcf_url}{Remote URL to VCF file.}

\item{vcf_dir}{Where to download VCF file.}

\item{download_method}{\code{"axel"} (multi-threaded) or
\code{"download.file"} (single-threaded) .}

\item{force_new}{Overwrite a previously downloaded VCF with the same path name.}

\item{quiet}{Run quietly.}

\item{nThread}{Number of threads to parallelize over.}
}
\value{
List containing the paths to the downloaded VCF and its index file.
}
\description{
Download a remote VCF file and its index file
}
\examples{
vcf_url <- "https://gwas.mrcieu.ac.uk/files/ieu-a-298/ieu-a-298.vcf.gz"
out_paths <- download_vcf(vcf_url = vcf_url)
}
