% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figshare_download.R
\name{figshare_download}
\alias{figshare_download}
\title{Download Figshare files}
\usage{
figshare_download(id, suffix = NULL, as_datatable = TRUE, verbose = TRUE)
}
\arguments{
\item{id}{Figshare repo ID.}

\item{suffix}{Filter results by file suffixes. 
Set to \code{NULL} to return all files (default).}

\item{as_datatable}{Return the return as 
a \link[data.table]{data.table} (default: \code{TRUE}),
instead of a nested list (\code{FALSE}).}

\item{verbose}{Print messages.}
}
\description{
Download files from a Figshare repository.
}
\examples{
files <- figshare_download(id="14703003", suffix="Earthworm_sFig1_cluster_marker.xlsx")
}
