% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figshare_list.R
\name{figshare_list}
\alias{figshare_list}
\title{List Figshare files}
\usage{
figshare_list(id, suffix = NULL, as_datatable = TRUE, verbose = TRUE)
}
\arguments{
\item{id}{Figshare repo ID.}

\item{suffix}{Filter results by file suffixes. 
Set to \code{NULL} to return all files (default).}

\item{as_datatable}{Return the return as 
a \link[data.table]{data.table} (default: \code{TRUE}),
instead of a nested list (\code{FALSE}).}

\item{verbose}{Print messages.}
}
\description{
List links to all downloadable files within a Figshare repository.
}
\examples{
files <- figshare_list(id="14703003", suffix=c(".RDS","cell_Info.txt"))
}
