% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zenodo_upload.R
\name{zenodo_upload}
\alias{zenodo_upload}
\title{Zenodo upload}
\source{
\code{
#### Installing zen4R ####
## has some system deps that have to be installed beforehand.
sudo apt-get install raptor2-utils
sudo apt-get install rasqal-utils
sudo apt-get install librdf0-dev
}
}
\usage{
zenodo_upload(
  files,
  token = Sys.getenv("zenodo_token"),
  title,
  zipfile = file.path(tempdir(), title),
  meta = list(title = title, description = "", uploadType = "dataset", Creator =
    list(firstname = "", lastname = "", affiliation = "")),
  sandbox = TRUE,
  validate = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{files}{Paths of files to upload to Zenodo.}

\item{token}{Zenodo
\href{https://zenodo.org/account/settings/applications/tokens/new/}{
Personal access token}.}

\item{title}{Metadata field: title of Zenodo repo.}

\item{zipfile}{Name of the zipped file containing all \code{files}.
This allows users to preview the folder hierarchy on the Zenodo website.}

\item{meta}{Additional Zenodo repo metadata.}

\item{sandbox}{Whether to use the Zenodo or Zenodo Sandbox API.}

\item{validate}{Whether to validate if each files was published.}

\item{verbose}{Print messages.}
}
\description{
Search for files and then upload them to Zenodo.
}
\examples{
 
files <- sapply(seq_len(3), function(i){
    tmp <- tempfile(fileext = ".txt")
    writeLines(letters, tmp)
    return(tmp)
})
\dontrun{
zout <- zenodo_upload(files=files,
                      title="test_repo")
}
}
