% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_vcf.R
\name{download_vcf}
\alias{download_vcf}
\title{Download a remote VCF file and its index file}
\usage{
download_vcf(
  vcf_url,
  vcf_dir = tempdir(),
  download_method = "download.file",
  force_new = FALSE,
  quiet = TRUE,
  nThread = parallel::detectCores() - 1
)
}
\arguments{
\item{vcf_url}{Remote URL to VCF file.}

\item{vcf_dir}{Where to download VCF file.}

\item{download_method}{\itemize{
\item{\code{"axel"} : }{Multi-threaded}
\item{\code{"wget"} : }{Single-threaded}
\item{\code{"download.file"} : }{Single-threaded}
\item{\code{"internal"} : }{Single-threaded 
(passed to \link[utils]{download.file})}
\item{\code{"wininet"} : }{Single-threaded 
(passed to \link[utils]{download.file})}
\item{\code{"libcurl"} : }{Single-threaded 
(passed to \link[utils]{download.file})}
\item{\code{"curl"} : }{Single-threaded 
(passed to \link[utils]{download.file})}
}}

\item{force_new}{Overwrite a previously downloaded VCF 
with the same path name.}

\item{quiet}{Run quietly.}

\item{nThread}{Number of threads to parallelize over.}
}
\value{
List containing the paths to the downloaded VCF and its index file.
}
\description{
Download a remote VCF file and its index file
}
\examples{
vcf_url <- "https://gwas.mrcieu.ac.uk/files/ieu-a-298/ieu-a-298.vcf.gz"
out_paths <- download_vcf(vcf_url = vcf_url)
}
