% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloader.R
\name{downloader}
\alias{downloader}
\title{downloader wrapper}
\usage{
downloader(
  input_url,
  output_path = file.path(tempdir(), basename(input_url)),
  download_method = c("axel", "wget", "download.file", "internal", "wininet",
    "libcurl", "wget", "curl"),
  background = FALSE,
  force_overwrite = FALSE,
  quiet = TRUE,
  show_progress = TRUE,
  continue = TRUE,
  nThread = parallel::detectCores() - 1,
  alternate = TRUE,
  check_certificates = TRUE,
  timeout = 30 * 60,
  conda_env = "echoR"
)
}
\arguments{
\item{input_url}{URL to remote file.}

\item{output_path}{The file name you want to save the download as.}

\item{download_method}{\itemize{
\item{\code{"axel"} : }{Multi-threaded}
\item{\code{"wget"} : }{Single-threaded}
\item{\code{"download.file"} : }{Single-threaded}
\item{\code{"internal"} : }{Single-threaded 
(passed to \link[utils]{download.file})}
\item{\code{"wininet"} : }{Single-threaded 
(passed to \link[utils]{download.file})}
\item{\code{"libcurl"} : }{Single-threaded 
(passed to \link[utils]{download.file})}
\item{\code{"curl"} : }{Single-threaded 
(passed to \link[utils]{download.file})}
}
 or
\code{"download.file"} (single-threaded) .}

\item{background}{Run in background}

\item{force_overwrite}{Overwrite existing file.}

\item{quiet}{Run quietly.}

\item{show_progress}{show_progress.}

\item{continue}{continue.}

\item{nThread}{Number of threads to parallelize over.}

\item{alternate}{alternate,}

\item{check_certificates}{check_certificates}

\item{timeout}{How many seconds before giving up on download.
Passed to \code{download.file}. Default: \code{30*60} (30min).}

\item{conda_env}{Conda environment to use.}
}
\value{
Local path to downloaded file.
}
\description{
R wrapper for \code{"axel"} (multi-threaded) and
 \code{"download.file"} (single-threaded) download functions.
}
\examples{
rda_url<-"https://github.com/RajLabMSSM/echolocatoR/raw/master/data/BST1.rda"
out_path <- downloadR::downloader(
    input_url = rda_url,
    download_method = "axel"
)
}
\seealso{
Other downloaders: 
\code{\link{axel}()},
\code{\link{wget}()}
}
\concept{downloaders}
