% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wget.R
\name{wget}
\alias{wget}
\title{wget}
\usage{
wget(
  input_url,
  output_path,
  wget_path = "wget",
  background = TRUE,
  force_overwrite = FALSE,
  quiet = FALSE,
  show_progress = TRUE,
  continue = TRUE,
  check_certificates = FALSE,
  conda_env = "echoR_mini",
  verbose = TRUE
)
}
\value{
Local path to downloaded file.
}
\description{
R wrapper for wget
}
\seealso{
Other downloaders: 
\code{\link{aws}()},
\code{\link{axel}()},
\code{\link{downloader}()}
}
\concept{downloaders}
\keyword{internal}
